Added a pause menu
